251125_날씨 메소드의 매개변수 타입 수정

+++ src/main/java/web/service/WeatherService.java
@@ -15,7 +15,7 @@ import java.time.format.DateTimeFormatter;
 @Service
 public class WeatherService {
     // 기상청 API를 활용하여 데이터 가져오기
-    public String getWeatherData(double lat, double lon) {
+    public String getWeatherData(int lat, int lon) {
         try {
             // 현재 시간 구하기
             LocalDateTime now = LocalDateTime.now();
